# guardian/services/api-gateway/openapi.yaml
openapi: 3.0.3
info:
  title: Guardian API
  description: API Gateway — auth, orders, bids, matches. Соответствует ABSOLUTE_STANDARD.
  version: 1.0.0

servers:
  - url: http://localhost:8080
    description: Local

paths:
  /health:
    get:
      summary: Health check
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok

  /api/v1/auth/register:
    post:
      summary: Register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, first_name, last_name]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 6 }
                first_name: { type: string }
                last_name: { type: string }
                user_type: { type: string, enum: [client, guard] }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: string, format: uuid }
                  token: { type: string }
                  user: { $ref: '#/components/schemas/User' }
        '400': { description: Bad request }
        '409': { description: Email already registered }

  /api/v1/auth/login:
    post:
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
                  user: { $ref: '#/components/schemas/User' }
        '401': { description: Invalid credentials }

  /api/v1/auth/me:
    get:
      summary: Current user
      security: [bearerAuth]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '401': { description: Auth required }

  /api/v1/orders:
    post:
      summary: Create order
      security: [bearerAuth]
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateOrderRequest' }
      responses:
        '201': { description: Created }
        '400': { description: Bad request }
        '401': { description: Auth required }
    get:
      summary: List my orders
      security: [bearerAuth]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  orders: { type: array, items: { $ref: '#/components/schemas/Order' } }

  /api/v1/orders/{id}:
    get:
      summary: Get order
      security: [bearerAuth]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: OK }
        '404': { description: Not found }
    patch:
      summary: Update order
      security: [bearerAuth]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: OK }
        '404': { description: Not found }

  /api/v1/orders/{id}/cancel:
    post:
      summary: Cancel order
      security: [bearerAuth]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: OK }
        '404': { description: Not found }

  /api/v1/bids:
    post:
      summary: Create bid
      security: [bearerAuth]
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateBidRequest' }
      responses:
        '201': { description: Created }
        '400': { description: Bad request }
        '401': { description: Auth required }

  /api/v1/my/bids:
    get:
      summary: List my bids
      security: [bearerAuth]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  bids: { type: array, items: { $ref: '#/components/schemas/Bid' } }

  /api/v1/bids/{id}:
    get:
      summary: Get bid
      security: [bearerAuth]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: OK }
        '404': { description: Not found }
    patch:
      summary: Update bid
      security: [bearerAuth]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: OK }
        '404': { description: Not found }

  /api/v1/matches:
    get:
      summary: List matches (for current user)
      security: [bearerAuth]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  matches: { type: array, items: { $ref: '#/components/schemas/Match' } }

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        first_name: { type: string }
        last_name: { type: string }
        user_type: { type: string }
        verified: { type: boolean }
        created_at: { type: string, format: date-time }
    Order:
      type: object
      properties:
        id: { type: string, format: uuid }
        client_id: { type: string, format: uuid }
        title: { type: string }
        description: { type: string }
        required_licenses: { type: array, items: { type: string } }
        budget_min: { type: number }
        budget_max: { type: number }
        latitude: { type: number }
        longitude: { type: number }
        start_time: { type: string, format: date-time }
        end_time: { type: string, format: date-time }
        status: { type: string }
        guard_count: { type: integer }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    Bid:
      type: object
      properties:
        id: { type: string, format: uuid }
        guard_id: { type: string, format: uuid }
        title: { type: string }
        licenses: { type: array, items: { type: string } }
        price_per_hour: { type: number }
        latitude: { type: number }
        longitude: { type: number }
        radius_km: { type: number }
        active: { type: boolean }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    Match:
      type: object
      properties:
        id: { type: string, format: uuid }
        order_id: { type: string, format: uuid }
        bid_id: { type: string, format: uuid }
        guard_id: { type: string, format: uuid }
        final_price: { type: number }
        created_at: { type: string, format: date-time }
    CreateOrderRequest:
      type: object
      required: [title, budget_min, budget_max, latitude, longitude, start_time, end_time]
      properties:
        title: { type: string }
        description: { type: string }
        required_licenses: { type: array, items: { type: string } }
        budget_min: { type: number }
        budget_max: { type: number }
        latitude: { type: number }
        longitude: { type: number }
        start_time: { type: string, format: date-time }
        end_time: { type: string, format: date-time }
        guard_count: { type: integer }
    CreateBidRequest:
      type: object
      required: [title, price_per_hour, latitude, longitude]
      properties:
        title: { type: string }
        licenses: { type: array, items: { type: string } }
        price_per_hour: { type: number }
        latitude: { type: number }
        longitude: { type: number }
        radius_km: { type: number }